+++ package.json
@@ -14,21 +14,18 @@
     "groq-sdk": "^0.7.0",
     "openai": "^4.72.0",
     "postcss": "^8.4.38",
+    "next": "^14.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-markdown": "^9.0.1",
-    "react-router": "^6.26.1",
-    "react-router-dom": "^6.26.1",
-    "react-scripts": "5.0.1",
     "remark-gfm": "^4.0.1",
     "tailwindcss": "^3.4.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
   },
   "eslintConfig": {
     "extends": [
