+++ pages/index.js
@@ -1,0 +1,5 @@
+import ConversationContainer from '../src/components/ConversationContainer';
+
+export default function Home() {
+  return <ConversationContainer />;
+}
