+++ pages/past-coversation.js
@@ -1,0 +1,5 @@
+import PastConversation from '../src/pages/PastConversation';
+
+export default function PastPage() {
+  return <PastConversation />;
+}
