+++ src/components/ConversationComp.js
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { PlusCircledIcon, MinusCircledIcon } from "@radix-ui/react-icons";
 import userIcon from "../assets/user-icon.png";
 import siteIcon from "../assets/site-icon.png";
-import { useLocation } from "react-router";
+import { useRouter } from "next/router";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 
@@ -28,8 +28,8 @@ function parseThink(text) {
 }
 
 function ConversationComp({ who, quesAns, time }) {
-  const location = useLocation();
-  const past = location.pathname === "/past-coversation";
+  const router = useRouter();
+  const past = router.pathname === "/past-coversation";
   const { reasoning, answer } = parseThink(quesAns);
   const style = past
     ? ""
